Reject duplicate peers and when we hit max peers
This isn't the most efficient way to do this, as it means that if we
receive a peer twice from two separte trackers, we'll still open two
connections. However, with this change the second connection will at
least be closed.

We also reject any connections that hit the maximum number of peers.
This only really applies to inbound connections though, as outbound
connections are checked against the max value when we receive them from
the tracker.